requirements.txt
Django==2.1.7
django-bootstrap-form==3.4
django-ckeditor==5.6.1
django-filter==2.1.0
django-js-asset==1.2.2
djangorestframework==3.9.2
Markdown==3.1
Pillow==5.4.1
pytz==2018.9
sorl-thumbnail==12.5.0
